* tree-gimple.c (is_gimple_addressable): Tidy. 	* Makefile.in (tree-ssa-structalias.o): Disable dependency on 	GTY header file. 	(tree-ssa-propagate.o): Likewise. 	(tree-phinodes.o): Likewise. 	(tree-scalar-evolution.o): Likewise. 	(tree-vect-generic.o): Likewise. 	* gimple.h (struct gimple_statement_phi): Change type of 	fields 'capacity' and 'nargs' to size_t. 	Update all users. 	* tree-cfg.c (verify_gimple_unary_expr): Remove.  Update all 	users. 	(verify_gimple_binary_expr): Remove.  Update all users. 	(verify_types_in_gimple_min_lval): Rename from 	verify_gimple_min_lval. 	(verify_types_in_gimple_reference): Rename from 	verify_gimple_reference. 	(verify_gimple_tree_expr): Remove. 	Move checks to verify_types_in_gimple_assign. 	(verify_types_in_gimple_call): Rename from verify_gimple_call. 	(verify_types_in_gimple_cond): Rename from verify_gimple_cond. 	(verify_types_in_gimple_assign): Rename from verify_gimple_assign. 	(verify_types_in_gimple_switch): Rename from verify_gimple_switch. 	(verify_types_in_gimple_phi): New. 	verify_types_in_gimple_return): Rename from 	verify_gimple_return. 	(verify_types_in_gimple_stmt): Rename from verify_gimple_stmt. 	(verify_types_in_gimple_seq): Rename from verify_gimple_seq.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gimple-tuples-branch@128010 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/gimple.c
@@ -466,10 +466,10 @@ build_gimple_try (gimple_seq eval, gimple_seq cleanup,
    ... phi_arg_d* for the incomming edges to this node.  */
 
 gimple
-build_gimple_phi (unsigned capacity, unsigned nargs, tree result, ...)
+build_gimple_phi (size_t capacity, size_t nargs, tree result, ...)
 {
   gimple p;
-  unsigned int i;
+  size_t i;
   va_list va;
   p = ggc_alloc_cleared (sizeof (struct gimple_statement_phi)
 			 + (sizeof (struct phi_arg_d) * (nargs - 1)) );
